refactor(overlay): add dedicated entry point
Move contextual positions, arrow calculation, and connected-overlay scroll strategies from common to the overlay entry point.

See #808

BREAKING CHANGE: move overlay utilities to the overlay entry point

Changed imports for positions, arrow calculation, and connected-overlay scroll strategies from @siemens/element-ng/common to @siemens/element-ng/overlay.
