+++ api-goldens/element-ng/overlay/index.api.md
@@ -1,0 +1,58 @@
+## Public API Report File for "@siemens/element-ng_overlay"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+
+import { ConnectedOverlayPositionChange } from '@angular/cdk/overlay';
+import { ConnectionPositionPair } from '@angular/cdk/overlay';
+import { ElementRef } from '@angular/core';
+import { ScrollStrategy } from '@angular/cdk/overlay';
+
+// @public
+export function calculateOverlayArrowPosition(change: ConnectedOverlayPositionChange, overlay: ElementRef, anchor?: ElementRef): OverlayArrowPosition;
+
+// @public
+export function defaultConnectedOverlayScrollStrategy(): ScrollStrategy;
+
+// @public
+export function defaultConnectedOverlayScrollStrategyFactory(): () => ScrollStrategy;
+
+// @public
+export interface OverlayArrowPosition {
+    // (undocumented)
+    left?: number;
+    // (undocumented)
+    right?: number;
+}
+
+// @public (undocumented)
+export const positionBottomCenter: ConnectionPositionPair;
+
+// @public (undocumented)
+export const positionBottomEnd: ConnectionPositionPair;
+
+// @public (undocumented)
+export const positionBottomStart: ConnectionPositionPair;
+
+// @public (undocumented)
+export const positions: {
+    top: ConnectionPositionPair[];
+    end: ConnectionPositionPair[];
+    bottom: ConnectionPositionPair[];
+    start: ConnectionPositionPair[];
+    auto: ConnectionPositionPair[];
+};
+
+// @public
+export const positionTopCenter: ConnectionPositionPair;
+
+// @public (undocumented)
+export const positionTopEnd: ConnectionPositionPair;
+
+// @public (undocumented)
+export const positionTopStart: ConnectionPositionPair;
+
+// (No @packageDocumentation comment for this package)
+
+```
