+++ projects/element-ng/common/models/index.ts
@@ -5,4 +5,3 @@
 export * from './color-variant.model';
 export * from './menu.model';
 export * from './status-type.model';
-export * from './positions.model';
