+++ projects/element-ng/date-range-filter/si-relative-date.component.spec.ts
@@ -5,7 +5,7 @@
 import { ScrollStrategy } from '@angular/cdk/overlay';
 import { Component, signal } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { defaultConnectedOverlayScrollStrategy } from '@siemens/element-ng/common';
+import { defaultConnectedOverlayScrollStrategy } from '@siemens/element-ng/overlay';
 import { t } from '@siemens/element-translate-ng/translate';
 
 import { SiRelativeDateComponent } from './si-relative-date.component';
