+++ projects/element-ng/overlay/index.ts
@@ -1,0 +1,7 @@
+/**
+ * Copyright (c) Siemens 2016 - 2026
+ * SPDX-License-Identifier: MIT
+ */
+export * from './connected-overlay-scroll-strategy';
+export * from './contextual-overlay-arrow';
+export * from './positions';
